// src/widget/types/home-search-params.tsx
export interface IHomeSearch {
  leasemiles: number;
  leasemonths: number;
  leasemoney: number;
  leaseregion: string;
  leasemaxcredit: number;
  carmake?: string;
  searchradius: number;
  leasetradein: boolean;
  leasemaxpayment: number;
  carincentive?: string;
  carmodel?: string;
  a?: string;
};